+++ apps/web/src/app/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { getDataSources } from './actions/get-data-sources';
-import { DataSourceManager } from './components/DataSourceManager';
+import { getDataSources } from '@/actions/get-data-sources';
+import { DataSourceManager } from '@/components/data-source/DataSourceManager';
 
 export default async function Home() {
   const dataSources = await getDataSources();
